requirements.txt
openai
Pillow
ollama